* win32/win32.c (rb_w32_fd_copy): cast explicitly to suppress warnings.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@33166 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ win32/win32.c
@@ -2470,8 +2470,8 @@ rb_w32_fdisset(int fd, fd_set *set)
 void
 rb_w32_fd_copy(rb_fdset_t *dst, const fd_set *src, int max)
 {
-    max = min(src->fd_count, max);
-    if ((UINT)dst->capa < max) {
+    max = min(src->fd_count, (UINT)max);
+    if ((UINT)dst->capa < (UINT)max) {
 	dst->capa = (src->fd_count / FD_SETSIZE + 1) * FD_SETSIZE;
 	dst->fdset = xrealloc(dst->fdset, sizeof(unsigned int) + sizeof(SOCKET) * dst->capa);
     }
